Remove autogenerated ANTLR files.
Preferably, autogenerated files should not be committed. It clutters the
repository and the diffs, and it means that you get multiple files that
carry basically the same information, one of which is not
human-readable. Also, it is not immediately clear whether they have been
kept in lockstep, as they should have been.

If we do want to distribute autogenerated files, we can do that by
providing a Release --- but that is separate from the repository, which
should remain clean with human-readable, plain-text *source* files.