Fix `graph2query` script.
I had switched around the 'before' and 'after'. But even after that, it
had some unused nodes, so honestly, even though I now just used the
queryEx structure that was already there, I'm not sure that it actually
works correctly: I think the fact that my previous script failed is
indicative of something.