Emphasize variability in corpus. (#7)
Knowing which parts of the text can be freeform and which should be
pinned down is a good first step towards making things more robust.

I'm actually not certain where variability occurs --- this is a first
quick guess. It seems way too broad to be dealt with by a parser
otherwise; I guarantee that the slightest variations will make it freak
out. Also, there are grammatical errors (which I've left in for now),
and I don't know how we can make the user understand how much freedom
they have in freeform text fields without requiring them to understand
what concepts they will recognize --- which is exactly what this is
supposed to abstract from.